Ensure at least SOME test also run on Tez
